fix: server crash from @shared/* runtime import for ManuscriptChat
Production crashed on boot with:
  Error [ERR_MODULE_NOT_FOUND]: Cannot find package '@shared/ManuscriptChat'
  imported from /app/server/dist/models/ManuscriptChat.js

Root cause: @shared/* is a tsconfig PATH ALIAS, resolved at compile time
only. The Heroku image copies shared/ as raw TypeScript (no build step
on the shared workspace), and Node has no resolver for the alias at
runtime. Existing files that use this pattern (e.g. models/Manuscript.ts
with `export * from '@shared/Manuscript'`) work only because the shared
file has zero runtime exports — tsc elides the runtime import in that
case.

shared/ManuscriptChat.ts shipped with `MANUSCRIPT_CHAT_MODELS` and
`DEFAULT_MANUSCRIPT_CHAT_MODEL` as runtime const exports, which forced
tsc to emit a real `import from '@shared/ManuscriptChat'` in the
compiled output. Production Node couldn't resolve it → boot crash.

Fix:
- shared/ManuscriptChat.ts is now types-only (matches the safer
  pattern used by shared/StoryCraft.ts), and its top-of-file comment
  explicitly warns future contributors not to add runtime exports.
- The runtime constants moved to a new server-only file
  server/src/services/manuscript-chat-models.ts.
- server/src/models/ManuscriptChat.ts now uses `export type *` so tsc
  emits no runtime import even if shared ever grows runtime exports
  again.
- The chat service and controller import constants from the new
  server-only file; types come from models/ManuscriptChat as before.
- The client never imported the constants — it already fetches the
  list via GET /api/manuscripts/chats/models — so no client changes.

Verified by inspecting compiled output: dist/server/src/models/
ManuscriptChat.js is now `export {};` and the controller imports the
constants from the local relative path.
